move time code to utils
